Split problem state into recovering and problem states
Dropping straight into problem when we encounter an error that might
be recoverable os a bit confusing, especially when we call that state
"Ready Retrying" in the UI.

To alleviate, create a new state named "recovering" that we transition
a node into when we detect an out of normal situation that might be
recoverable, and reserve "problem" for when the recovery steps we might
want to try have failed to fix things.

In the UI:
problem: Problem <red flasher>
recovering Recovering <red flasher>

There is no more "Ready Retrying" anymore.