Added SUSE support to apache cookbook
This adapts the apache cookbook to create a apache configuration on SLES good
enough for getting openstack-horizon dashboard running. Using SUSE's
default apache configuration layout, instead of adapting it to Debians/Ubuntu's
layout.

TODO: This is by no means complete and does not support all of the features
that this cookbook supports on Ubuntu.