Add SUSE init script for crowbar (still needs work)
This invokes rainbows directly, it doesn't use bluepill (this is
intentional).  The stop and status functions don't currently work,
and a pidfile is not saved.  I suspect this is due to some
discrepancy between the init script specifying /usr/bin/rainbows,
but the running process only appearing as 'rainbows'.

Signed-off-by: Tim Serong <tserong@suse.com>

Conflicts:
	chef/cookbooks/crowbar/recipes/default.rb