Switch back to platform check for a2enflag call (bnc#775998)
Using File.exist?("/etc/sysconfig/apache2") doesn't really work, as it is
executed in the compile phase of chef. Long before the apache2 package is
actually installed by the apache2 recipe, so it's likely that
/etc/sysconfig/apache2 doesn't exist at that stage.

Addtionally we use the "execute" resource instead of calling system to execute
a2enflag now, which is a lot cleaner.