Add a third validation steps for proposal, after saving
With the current validation of proposals, all changes to the proposals
are lost. This is good for changes that are really destructive, but
annoying for changes that are easy to fix.

Therefore, this patch adds a validation step after the saving of the
proposal. This will be up to each barclamp to override this method.

https://bugzilla.novell.com/show_bug.cgi?id=776375