Fix NodeObject.find_node_by_alias
The query it was doing was just for any alias field. It turns out that
there are alias fields for nodes with the dns-server role. And we don't
want to hit those values.

This is highly visible when trying to give an alias to a node that it
had previously, since the dns-server fields are not updated immediately
and might still have the old value. This will lead to a conflict.

So instead, search for the crowbar/display/alias nested field (which can
be done by searching for "crowbar_display_alias", see
http://wiki.opscode.com/display/chef/Search)