Replace recursive_merge! with Chef's deep_merge!
recursive_merge! had a number of serious issues, worst of which was
that it did not recurse.

An exclamation mark was missing in the recursive call.  However the
interpreter never even reached this typo in order to result in a
NoMethodError, because the block passed to merge! was never used - b
is always a Mash and unlike Ruby's Hash, Chef's Mash ignores any block
passed to merge!  So in fact no recursion ever took place.

Other issues:

    if _old.class.kind_of? Hash.class

should have read

    if _old.kind_of? Hash

There should have also been a test when recursing that _new was a kind
of Hash.

In the case where _new was nil, it should not have overridden the
existing value.

https://bugzilla.novell.com/show_bug.cgi?id=775654

(cherry picked from commit 385467bb694ef1548cfb2213a4da58b14700912a)