First draft of the proposed browser support policy
A clearly stated browser support policy makes it clear to our users what
to expect, and which browser and OS combination they should use for the
optimal experience.

It also makes it clear to developers and QA what the priorities are.
This way we can focus the limited resources on the tasks that matter
most.