[Ruby 1.9] Fix bogus asserts in barclamp_model unit test
`assert` is incorrectly used here, since it just tests that the first
argument is `true`, which in Ruby means _any_ object (eg. 0, 1, "", [],
{}) except `false` and `nil`.

Ruby 1.9 is stricter here and checks that the second argument is a
`String` or `Proc`, thus uncovering the bug/typo. `assert_equal` should
be used here instead.

Also ensure that the order of the two arguments of `assert_equal` is
correct. That is, expected result first followed by actual. This
prevents confusion when reading the error messages in failing tests.

Finally, prefer using single quotes instead of double quotes when string
interpolation is not needed.