Change how we calculate fingerprints.
Saving the individual fingerprints for a node and then adding was
giving me intermittent ineger overflows in the database.  Since
fingerprint is only used to force a refresh of the UI when we add or
remove a node, instead add a class method to the Node model that
returns the sha1sum of all the node names sorted in ascending order.
This achieves the same goal without caring about having to worry about
how Ruby string hash values will fit into the database or worrying
about whiich value could trigger an integer overflow in the database.