Switch around special-case hacking to make the admin node install more regular.
Instead of special-casing how we make the machine-install user, switch up the
install process to let us create the machine-install user and the admin node
using regular CLI and API calls.

This involves special-casing the authentication paths in the framework to skip
auth checking when the framework is being bootstrapped and we are making API
calls from localhost.