Use Settings.doc.debug to determine docs debug output
This gives us more fine grained control on what we see in development
mode, and not get forced to see the debugging output when we don't want
to.

The default is set to 'true' to maintain the same behavior as before. It
can be overridden by copying settings.yml to settings.local.yml, and
then changing it to 'false'.