run_in_prod is no longer needed.
We only have a Chef jig if one was created via the inject_chef_conn
rake task, so we no longer need this guard for the unit test.

While we are at it, get rid of a redundant call to Jig.delete_node in the
node model -- for now, we handle calling the jigs for create and delete in
the controller.