Make docs rebuilding configurable
The docs are hardcoded to rebuild when in development mode, unless the
`rebuild` parameter is passed in. This makes it very slow (>10s) when
working on it, so let's just make it configurable.

Rebuild still on by default when in development mode, and off in
production to maintain the same behavior. But it can now be easily
configured locally in `config/settings.local.yml`.