Improve UI of list of barclamps
When only one proposal is allowed, do not show an expander which
shows/hides the list of proposals: directly put the Create/Edit button
in the barclamp row.

Also, remove links that show a proposal in non-edit mode: this page
displays the content of the json, which is not really helpful. Always go
to edit mode, which is what people are interested in.