Add render_instance_selector helper to render instance selector
It works like the instance selectors in templates, except that it also
checks if multiple proposals are allowed for a barclamp, and doesn't
show anything if this is not the case.

This will make it easy to change, temporarily or not, the
allow_multiple_proposals of barclamps. And this avoids repeating these
three template lines several times in each barclamp template.