webui: Remove cancel button on edit proposal page
There is actually no cancel action: by clicking cancel, you don't cancel
anything, you just leave the page. There's no need for a button
dedicated to this (since there's the whole navigation menu which can be
used).

This was also adding confusion, as people might have thought that cancel
could be used to deactivate a proposal ("I want to cancel this
proposal"), or to cancel the creation of a proposal just after it got
created.