DE1262: Fix rabbitmq-server going away.
When the logs are rotated, bluepill chef-server restart winds up
getting restarted several times in quick succession.  This can cause
bluepill to get trapped in a state where it thinks Rabbit is up when
it is not.

To work around this, have logrotate make a copy of the current log and
then truncate it.  This gets rid of the need to have the services
restart in order to continue logging messages.