[Bundler] Disable use of Bundler by default
There are many concerns that the re-introduction of Bundler, especially
at this stage of the release, will cause more problems than benefits.
But it is still very useful to have the ability to use Bundler,
particularly during development. So let's take the middle ground and
push the code in, but disable it by default.

See the discussion at
https://github.com/crowbar/barclamp-crowbar/pull/557 for details.

* Use the `app_config` gem [1] for configuring the use of Bundler (off
  disabled by default). Will be used for other config options as well.
  It's similar to the `rails_config` gem [2] that's already used in
  Crowbar 2.x, but works for Rails 2 (rails_config doesn't).

* Gemfile and Gemfile.lock are ignored by default, and only parsed when
  `use_bundler` is set to `true` (which is very useful during
  development).

[1] https://github.com/Oshuma/app_config/tree/v1.0.2
[2] https://github.com/railsjedi/rails_config