Make jig even more late-bound.
This allows us to load all of the roles a barclamp has even though
we may not yet have the jigs created to handle them.

It also changes how we try to load roles in a crowbar.yml to make
the process less jig-oriented.  The eventual goal there is to have the
jigs section concern itself with declaring what jigs are provided by a
barclamp.