don't add deleted nodes to the run order (bnc#840255)
If a node is deleted but still associated with a proposal,
and then it is removed from that proposal and the proposal
applied, we need to avoid attempting to adjust the run list
for that node, or attempting to run chef-client on that node,
or queuing actions.

Previously this would result in a crash such as:

  undefined method `[]’ for nil:NilClass

leaving all nodes stuck in the applying state, and
the proposal stuck in the committing state.

  https://bugzilla.novell.com/show_bug.cgi?id=840255

See also

  https://bugzilla.novell.com/show_bug.cgi?id=839714