ServiceObject#apply_role: rename nodes to pending_node_actions
ServiceObject#apply_role is a grossly overgrown and complicated method,
and also just happens to be probably the most critical part of Crowbar's
core engine.  So anything we can do to make it more legible is a
worthwhile investment.  Renaming and explaining cryptic variables helps,
as does moving variables closer to where they are used.

This is a refactoring only, and should not affect behaviour in any way.