rename cryptic "r_nodes" variable to "nodes_in_batch"
ServiceObject#apply_role is a grossly overgrown and complicated
method, and also just happens to be probably the most critical
part of Crowbar's core engine.  So anything we can do to make
it more legible is a worthwhile investment.

This is a refactoring only, and should not affect behaviour in any way.