Replace Marshal dump+load w/ deep clone method, refactor calls
Marshal can unserialize specially crafted strings to a class, on which the user
might attempt to call arbitrary methods.

This should be a false positive, as we call only methods defined by us on the
cloned object and the unserialized value itself is probably not user supplied.

Anyway, use a deep_copy method that won't contain the special unserializing
magic.