Add a model to track user, system, and wall data for noderoles.
This allows us to relax the coupling between noderoles and snapshots,
which reduces the overall complexity of the noderole graph.

Before this change, creating a new snapshot for a deployemnt would cause
us to duplicate sections of the noderole graph, leading to an
accumulation of garbage in the node_roles and node_role_pcms tables that
we then had to have logic to skip when we wanted to perform calculations
on the noderole graph.  The reason we did this was to track changes in
user data over time.

This change extracts the userdata, systemdata, and wall fields of the
node_roles table into their own table, and fixes up the noderole logic
to refer to the new table.  The new table is curently treated as
append-only -- any change to a noderole's wall, sysdata, or user data
saves a new record with the updated data and points the noderole at that
one instead.