Remove references to CHEF_ONLINE from the code
This constant was used previously for testing. However, we currently use
OfflineChef (spec/support/offline_chef.rb) to stub chef responses, so
this is no longer needed.

This patch takes care of removing the code dependent on this constant
being false, i.e. we assume that CHEF_ONLINE is always true. This
simplifies code logic on a lot of places.