Improve suggested proposal name on creation so that it's unique
Until now, the default proposal name is "proposal" for all barclamps.
This is fine when the barclamp does not support multiple proposals, but
when it does, it means that the suggested proposal name might already be
in use.

Now we append a _1, _2, etc. string to the suggestion for the new
proposal name. That way, we guarantee that the user can directly create
the proposal.

(cherry picked from commit a3ecc894e38078b5c9253f75fcabfe4192b5f559)