Plugin for toggling visibility on cluster deploy
We need to show/hide certain parts of barclamp configuration only when a
cluster is deployed. This jQuery plugin adds that functionality.

Example usage:

<fieldset data-elements-path="database-server" data-show-for-clusters-only="true">

This element (and its children) will be shown only when:

1. A cluster is already part of a deployment
2. A cluster is dragged into the node list

The data-elements-path specifies where to check in the deployment JSON
for already deployed clusters. data-show-for-clusters-only enables this
plugin for a given element.

(cherry picked from commit 1b710e3121128237ac555e385a23534113d421d3)