Handle removed node referenced by another proposal
When another proposal removes a node which the current proposal
references, the corresponding DOM element is no longer available in the
deployment list. However, info from this element is needed by the
nodeList, when it triggers an event that a node has been removed.

Make the nodeList look up the info in some other way.