dynamicTable overwrites changes w/ old data
BNC#865526

The dynamicTable plugin caches the proposal attributes JSON in it's json
property and does not take into account that the data might have
changed when it writes its own copy back.

This fix minimizes the race window. Another one would be to
make sure we always operate only on the path + value that changed.