fix broken unclaimed_physical_drives method
The list returned by `phyisical_drives` contains two element lists where
the first element is the OS mount path/name for the device. This then
needs to be turned into the device unique name in order to use the
`disk_owner` method.

This also fixes the unclaimed physical disks validation in the cinder
barclamp.

Fixes bnc#830868