Display node list in node show view as well
https://bugzilla.novell.com/show_bug.cgi?id=872347#c5

Once in the node show view, there is no easy way to navigate to other
node w/o going to the dashboard.

Display the (uneditable) node list below the node info.

Refactor the fetching of node groups into a separate method, then re-use
it in the show action. Similarly, split out the node list table into a
partial, then render the partial in both the index and the show views.