Bulk edit node alias exchange (bnc#877090)
The ticket https://bugzilla.novell.com/show_bug.cgi?id=877090 explains
that it is not possible to switch the names for two or more nodes on the
bulk edit page, it results in an error where it simply sais failed to
save the nodes.

I have split the ```alias=``` and ```public_name=``` methods into
multiple methods to integrate a force setter method to avoid the
duplication check because we are not doing real bulk editing.

As the duplication check is already included in the controller for bulk
editing we don't need to check that again within the setter.

To show the user a usefull error message i have added one more
translated error message for the public name as well.