Mark proposals as applied on commit
This is another attempt for a fix of
https://bugzilla.novell.com/show_bug.cgi?id=861309

The previous solution was based on comparing the crowbar-revision of
proposal and the corresponding role. However, these numbers can diverge
arbitrarily and the comparison would not work as expected, as was observed in
https://bugzilla.novell.com/show_bug.cgi?id=877486

This patch uses a dedicated 'committed' attribute, which defaults to
false. It is set to 'true' after a proposal is committed and re-set back
to 'false' on any other operation.