Fix UI if chef-server is gone (bnc#865072)
The ticket https://bugzilla.novell.com/show_bug.cgi?id=865072 explains
that crowbar behaves bad if the chef-server is gone and not reachable.

In order to display a better usable error page i have integrated more
exception classes that gets catched by the application controller to
display a proper error page if there are documents not found or the chef
server is gone.