Added more strict gem version constraint
I have added a more strict version constraint because the previous
versioning doesn't work if you work on the git base. If you bundle
install in the crowbar_framework folder it always installs upstream chef
10.32.x and ohai 6.22.x. Installing this chef version breaks pacemaker
as it works with subfolders on libraries folders, these files won't get
uploaded to the chef server with upstream chef client.