Fix race with multiple chef-client runs
Handle possible chef-client runs triggered by crowbar when the target
node is currently rebooting. In This case, just wait until the reboot
sucessfully finished and trigger the run after the reboot.
This can happen if a chef-client run (triggered by the chef-client
daemon) triggers a reboot and after that the chef-client (triggered by
crowbar) starts.