Support finding crowbar subcommands when the executable is a symlink
Right now, the crowbar tool only looks for subcommands in the current
directory and in the directory of the executable. We change the latter
to look for subcommand in the directory of the symlink target if the
executable is a symlink.

This allows having a /usr/bin/crowbar symlink pointing to
/opt/dell/bin/crowbar.