Drop gather_cli.sh script
The idea was to allow exporting the cli tools from the rails app, but
this code in the rails app is dead. And the generation of the tarball is
kind of broken anyway (very Ubuntu-specific).

If we want to allow distribution of the cli tools, we should do it
another way (separate git repo, for instance).