Controllers: Preload chef objects in the controller so they are not reloaded in views
The dashboard suffers the same problem as every other - large amounts of
data needs to be pulled in. However, with chef (api), the round trip of
one request is up to 30ms, so loading anything repeatedly or fetching
single records is out of the question.

So we do two things:

1) Preload all *Objects. This is inefficient, but still is faster than
fetching individual records or running find_all_by queries.

2) Methods that search for certain kinds of *Objects or fetch associated
records of an *Object accept an array of unfiltered values retrieved in
1), then operate on that, rather then querying on their own.