Don't allocate node on proposal apply (bnc#863101)
Applying a barclamp allocates (i.e. re-installs) a node.  This could be
problematic e.g. if the node contains a previously installed OS, or user
data.

This patch changes the policy to not allocate the nodes automatically on
barclamp apply.

This is preferrable to not offering the non-allocated nodes in the UI as
it works also for the API. User will be informed about the need to
allocate the nodes manually.