Do not use deep munge
This is breaking the JSON parsing for empty arrays, which we use with
the magic crowbar-deep-merge-template feature when creating the initial
crowbar proposal.

With deep munge enabled, we cannot remove the nagios proposal, which
results in failures when nagios barclamp is not available.

See:
https://github.com/rails/rails/issues/8832
https://github.com/rails/rails/issues/13420

We can revert this once switching to rails 5 (see comments in second
issue).