Load Rack::ContentLength and Rack::Chunked
Explicitly use Rack::ContentLength and Rack::Chunked in the middleware
stack.

Whenever a response is missing either Content-Length or Chunked
encoding header, there is a few seconds lag before Rainbows closes
the connection to indicate the end of the response body [1].

Rainbows includes the necessary rack middleware by default, but only for
'deployment and development' settings of RACK_ENV env variable, i.e. not
for production. So for production, it needs to be included manually.

Registering the modules via Rails.configuration.middleware.use breaks
mini-profiler, so they're added to the rackup file directly.

[1] http://www.w3.org/Protocols/HTTP/1.1/draft-ietf-http-v11-spec-01.html#BodyLength