Try protected methods in attr DSL (bnc#912426)
The cinder backend UI uses a hack to workaround handlebars substituted
values and calls wrap_around - a method defined in form helper.

The helpers possibly might have been public methods of the view object
in rails 2, but they're protected in rails 4.

A respond_to?(method) check on the view object then returned false
and NoMethdodError was raised.

Fix the respond to check to search for protected and private methods as
well.