Node list and led updates on change (bnc#912426)
Whenever a node info changes or a node is re-added into a different
group, make sure the UI reloads the node page and does not redirect
to the dashboard.

The only case when we really want to go to the dashboard instead is
when we 'forget' the node and this is handled by the previous patch.