Move the list of valid states to controller
We do not provide validations or post/pre condition hooks for
transition and barclamps (e.g. ntp, deployer...) do not call
super, so we can't add the check to the superclass either.

Easiest way how to check the state is add it to the controller. This is
definitely not optimal, though.