Totally ignore nodes that were never discovered by crowbar (bnc#900216)
If chef-client is run on a node, but that node never went through the
discovered state, then crowbar never created the role for that node
(crowbar-$node).

This will result in failures when we list nodes because chef knows about
the node, and crowbar thinks it should know about it too.

Instead, if we see that the node role doesn't exist, then we pretend the
node just doesn't even exist.

https://bugzilla.suse.com/show_bug.cgi?id=900216