Fix alias in bulk edit (bnc#918523)
If you try to set an alias or public_name on an unallocated node in bulk
edit, and you try to allocate it at the same time crowbar ignores the
alias and public_name you defined. Because of that i have moved the
alias and public_name assignment below the allocation, that way this
information doesn't get lost.

https://bugzilla.suse.com/show_bug.cgi?id=918523