Allow passing in logger as an option to RemoteNode methods
RemoteNode uses puts as a default way of logging (since it assumes it
will be called from a forked process, where all file descriptors except
for stdout/in/err are closed).

However, it now needs to be used in another context, namely to check if
chef clients are running prior to applying of the role.

This patch adds an options hash to relevant methods and allows passing a
logger option. If not specified, it falls back to using puts.