Add proposal versions
Do not merge (the feature currently does not work due to proposal
having overrided [] method). Also, code will be separated into
smaller commits and covered by tests.

The table holds history of a given proposal. The records are full
copies and can be used to restore the original model.

State during create, update and destroy actions is recorded.

A deployment task will not apply the proposal at its current state,
but its version when 'apply' was originally clicked.

This allows the user to modify the proposal afterwards, or even delete
it, and the deployment queue will not or apply unexpected changes, or
get into an inconsistent state, as it will hold the complete info.

Additionally, UX tricks like 'undo', 'view history' or 'view cloud state
at past deployment', etc. are possible.

There are a few unknowns:

1. Should this 'snapshotting' extend to the proposal dependencies or
even all proposals as well? The deploy would then check that all the
dependent proposals have been applied (and do not differ), then
selectively apply them?

2. When a proposal is deleted, should we enqueue another 'cleanup'
deploy, that gives the cookbooks a chance to clean up?