Mark nodes as applying a bit later in the code path
We were doing it too early, outside of apply_role, which meant losing
some information like the fact that nodes that are running a "remove
role" chef-client run are also applying.

Now we instead use the list of nodes we will act on.

Also use this list when restoring the state of nodes to ready.