Tighten up Crowbar::Product::is_ses? check
Previously this would just check if suse_enterprise_storage was in the
BarclampCatalog.  Unfortunately, this will always be true if the Ceph
barclamp is installed, as it's a member of suse_enterprise_storage since:

  https://github.com/crowbar/barclamp-ceph/pull/154

See subsequent discussion at:

  https://github.com/crowbar/barclamp-crowbar/pull/1266

This fix tightents up the check to verify that suse_enterprise_storage
is a member of itself, which won't happen unless that barclamp is
really, truly installed.

Signed-off-by: Tim Serong <tserong@suse.com>