prevent crowbar batch from locking itself and the user out
If crowbar batch is used to export the crowbar proposal from one cloud
and apply it to another, it will change the machine-install user's
password, which is typically the one used to run crowbar batch.  This
would break any future HTTP digest authentications against the Crowbar
REST API, including ones invoked immediately after the change, by the
still running crowbar batch process.  So if the user inadvertently
attempts to change the password they're currently relying on, simply
output a warning and ignore the attempt.