rewrite Crowbar::Lock as a proper class
Previously this class only contained class methods, so it was
effectively a badly-written module.  Convert it into a real
class where lock objects can be instantiated and then #acquire
and #lock methods invoked on them.

Also add a #with_lock method which accepts a block so that critical
regions can be cleanly wrapped inside a lock.