add Lock::SharedNonBlocking class
This class implements a shared (i.e. non-exclusive) lock, by default on
a remote node.  It will be used by the next commit to allow one or more
concurrent proposal applications to temporarily pause any intervallic
runs triggered by the chef-client daemon, to prevent them from
interfering with the orchestration of the proposals being applied:

  https://bugzilla.suse.com/show_bug.cgi?id=857375

It has the same semantics as a shared read/write lock, except that
instead of readers reading, we have threads applying proposals, and
instead of a single writer writing, we have the chef-client daemon
triggering a chef-client run.  In other words, the daemon can only
trigger a run when there are no proposals being applied.