To to fix the provider network parameters of the floating network (bsc#967009)
Because of several bugs (at least bsc#946882 and bsc#955811) we have setups
where the floating network was created with the wrong provider network
settings. While this somehow continued to work and got unnoticed it will create
problems on system upgrading from juno to liberty. Try to adjust the parameters
directly in the database as part of the upgrade.

https://bugzilla.suse.com/show_bug.cgi?id=967009